Add test and refactor
